Move \Commands to \Console
